+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Menu that aggregates all download buttons
+
 ### Fixed
+
+- Excel padding cells not formatted
 
 ### Removed