+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Menu that aggregates all download buttons
 - Text to change the color of each course in the chip
 - Change text color of lessons to contrast well with the background color
+- Add missing campus on some lessons and improve formatting
 
 ### Fixed